misc fixes

+++ champions/src/main/java/me/mykindos/betterpvp/champions/champions/skills/skills/brute/sword/WhirlwindSword.java
@@ -121,7 +121,7 @@ public class WhirlwindSword extends Skill implements InteractSkill, CooldownSkil
                 }
 
                 for (int p = 0; p < particlesPerTick; p++) {
-                    double progress = (i + (double)p * increment / particlesPerTick) % points;
+                    double progress = (i + p * increment / particlesPerTick) % points;
                     for (int spiral = 0; spiral < numSpirals; spiral++) {
                         double startAngle = spiral * Math.PI * 2 / numSpirals;
                         double angle = 2 * Math.PI * progress / points + startAngle;
